fix package files directly
this would (hopefully) fix the issue of the invalid characters, or we can do a release.
the changes are - inside each tarball there is a package.json.

Cause: Probably i copied the wrong file while doing a local build. I did a local release build because the IG was too big to build via github actions.
How  to check: open the tarballs. the file package.json should have url
  "url" : "http://smart.who.int/ddcc/1.0.0",
and not
  "url" : "file://C:\\work\\ImplementationGuides\\ig-release\\who\\ddcc\\output",

the line
  "notForPublication" : true,
was also removed

how to test - after this merge, the package should be picked up by the tools. if not, we will need to make a technical release. I am waiting for a fix for that. so if that comes quickly we may not need to do this update